# components/manager/api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a swagger documentation for Reporter",
        "title": "Reporter",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:4005",
    "basePath": "/",
    "paths": {
        "/v1/data-sources": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all data sources connected on plugin with all information from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source"
                ],
                "summary": "Get all data sources connected on reporter",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/DataSourceInformation"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
        "/v1/data-sources/{dataSourceId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a data sources information with data source id passed",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data source"
                ],
                "summary": "Get a data sources information",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Data source ID",
                        "name": "dataSourceId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DataSourceInformation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "DataSourceInformation": {
            "description": "DataSourceDetails is the data source details of database",
            "type": "object",
            "properties": {
                "externalName": {
                    "type": "string",
                    "example": "onboarding"
                },
                "id": {
                    "type": "string",
                    "example": "midaz_onboarding"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TableDetails"
                    }
                },
                "type": {
                    "type": "string",
                    "example": "postgresql"
                }
            }
        },
...
        "TableDetails": {
            "description": "TableDetails is the struct of table information",
            "type": "object",
            "properties": {
                "fields": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "['id'",
                        " 'name'",
                        " 'parent_account_id']"
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "account"
                }
            }
        },
...
        "pkg.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "entityType": {
                    "type": "string"
                },
                "err": {},
                "message": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "The authorization token in the 'Bearer access_token' format. Only required when auth plugin is enabled.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}